pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "sofascore-mini"
version = "0.1.0"
description = "Mini SofaScore-like live football service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
